+++ frykit/__init__.py
@@ -1,5 +1,5 @@
 from pathlib import PurePath
 
-__version__ = '0.3.3'
+__version__ = '0.3.4'
 
 DATA_DIRPATH = PurePath(__file__).parent / 'data'